+++ tests/e2e/api/token.test.ts
@@ -40,11 +40,12 @@ async function setupToken(): Promise<string> {
     name: tokenName,
     uri: tokenUri,
   });
-  const response = await aptos.signAndSubmitTransaction({ signer: creator, transaction });
-  await waitForTransaction({ aptosConfig: config, transactionHash: response.hash });
+  const pendingTxn = await aptos.signAndSubmitTransaction({ signer: creator, transaction });
+  const response = await waitForTransaction({ aptosConfig: config, transactionHash: pendingTxn.hash });
   return (
     await aptos.getOwnedTokens({
       ownerAddress: creator.accountAddress.toString(),
+      minimumLedgerVersion: response.version,
     })
   )[0].current_token_data?.token_data_id!;
 }